# swagger.yaml
openapi: 3.0.3
info:
  title: RoutePilot API
  description: |
    Mock API for RoutePilot - a ridesharing platform connecting drivers and customers.

    This API supports real-time location tracking, route management, booking system,
    and in-app messaging between drivers and customers.

    **Features:**
    - User authentication (register, login, logout)
    - Route creation and management (drivers)
    - Booking system (customers)
    - Real-time location tracking
    - In-app messaging system
    - Statistics and analytics
    - WebSocket support for real-time updates
  version: 1.0.0
  contact:
    name: RoutePilot Support
    email: support@routepilot.com
servers:
  - url: http://localhost:5000/api
    description: Local development server
  - url: http://localhost:5000
    description: WebSocket server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Session token obtained from login/register
  schemas:
...
    Route:
      type: object
      properties:
        id:
          type: string
          format: uuid
        driverId:
          type: string
          format: uuid
        driverName:
          type: string
        driverAvatar:
          type: string
          format: uri
        driverRating:
          type: number
          format: float
        routeName:
          type: string
        origin:
          type: string
        destination:
          type: string
        departureDate:
          type: string
          format: date
        departureTime:
          type: string
          format: time
        stops:
          type: array
          items:
            $ref: '#/components/schemas/RouteStop'
        totalSeats:
          type: integer
          minimum: 1
          maximum: 8
        availableSeats:
          type: integer
        price:
          type: number
          format: float
        notes:
          type: string
        status:
          type: string
          enum: [active, cancelled, completed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RouteStop:
      type: object
      properties:
        id:
          type: string
          format: uuid
        routeId:
          type: string
          format: uuid
        name:
          type: string
        estimatedTime:
          type: string
          format: time
        order:
          type: integer
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        details:
          type: array
          items:
            type: object
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
paths:
...
  /routes:
    get:
      tags:
        - Routes
      summary: Search and list routes
      description: Get list of routes with optional filters and pagination
      parameters:
        - name: origin
          in: query
          schema:
            type: string
        - name: destination
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: minSeats
          in: query
          schema:
            type: integer
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: driverId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Routes retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Route'
                  pagination:
                    $ref: '#/components/schemas/PaginationInfo'

    post:
      tags:
        - Routes
      summary: Create new route
      description: Create a new route (drivers only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - routeName
                - origin
                - destination
                - departureDate
                - departureTime
                - totalSeats
                - price
                - stops
              properties:
                routeName:
                  type: string
                origin:
                  type: string
                destination:
                  type: string
                departureDate:
                  type: string
                  format: date
                departureTime:
                  type: string
                  format: time
                totalSeats:
                  type: integer
                  minimum: 1
                  maximum: 8
                price:
                  type: number
                  minimum: 0
                stops:
                  type: array
                  minItems: 2
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      estimatedTime:
                        type: string
                      order:
                        type: integer
                      latitude:
                        type: number
                      longitude:
                        type: number
                notes:
                  type: string
      responses:
        '201':
          description: Route created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Route'
        '403':
          description: Only drivers can create routes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /routes/{id}:
    get:
      tags:
        - Routes
      summary: Get route details
      description: Retrieve detailed information about a specific route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Route details retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Route'
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Routes
      summary: Update route
      description: Update route information (owner only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                routeName:
                  type: string
                departureDate:
                  type: string
                departureTime:
                  type: string
                totalSeats:
                  type: integer
                price:
                  type: number
                notes:
                  type: string
                status:
                  type: string
                  enum: [active, cancelled, completed]
      responses:
        '200':
          description: Route updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Route'
        '403':
          description: Not authorized to update this route
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Routes
      summary: Delete route
      description: Delete a route (owner only, no active bookings)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Route deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Cannot delete route with active bookings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...